### 2018-05-30 #759 @whedon
[PRE REVIEW]: Salabim: discrete event simulation and animation in Python
**Submitting author:** @salabim (<a href="http://orcid.org/0000-0001-7696-8059">Ruud van der Ham</a>)
**Repository:** <a href="https://github.com/salabim/salabim" target ="_blank">https://github.com/salabim/salabim</a>
**Version:** 2.2.23
**Reviewers:** @gonsie, pspringer, @pspringer
