### 2018-06-04 #759 @danielskatz
@salabim - do you have any suggestions for reviewers?  Perhaps from [this list of people](https://bit.ly/joss-reviewers) who have already agreed to review for JOSS and may be suitable for this submission?
