### 2018-06-06 #767 @whedon
[REVIEW]: Salabim: discrete event simulation and animation in Python
**Submitting author:** @salabim (<a href="http://orcid.org/0000-0001-7696-8059">Ruud van der Ham</a>)
**Repository:** <a href="https://github.com/salabim/salabim" target ="_blank">https://github.com/salabim/salabim</a>
**Version:** 2.2.23
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.6791129" target="_blank">10.6084/m9.figshare.6791129</a>
