### 2018-06-06 #767 @danielskatz
👋  @arfon - I screwed up and added @pspringer twice as a reviewer.  I'm going to manually edit the first comment in the issue to fix this, but if there's something in the @whedon database that also needs to be fixed, please do so.
