### 2018-06-08 #767 @pspringer
Hi Dan, this is Paul.  I've not done this sort of collaboration before, so I am completely lost.  I don't even know if you will get this if I don't put that @ sign in front of your name.  I'm not even sure what button to click at the bottom to send this.  At any rate, if you could point me to a guide for how to do this, it would help.  Thanks
