### 2018-06-08 #767 @danielskatz
@pspringer - We're using GitHub's notification feature within issues, so if you put @danielskatz, I will be notified.

And these are typical GitHub issues - fully public for anyone who cares to look, and continually added to as more comments are added (by the green comment button under the empty box at the end of a current list of comments on an issue)

It might be useful for us to talk offline on some of the details of how this reviewing process works within GitHub, and not bother others who may be following this issue.  I'll followup by email.
