### 2018-06-08 salabim/salabim#2 @gonsie
Tagged release
Could you add a git or GitHub tag for your latest release (version 2.2.23)? It can be done on this page: https://github.com/salabim/salabim/releases

The tagged release is needed for openjournals/joss-reviews#767
