### 2018-06-08 salabim/salabim#3 @gonsie
Repository organization
Hi,

The files in this respository need to be organized. A typical repository has the following directories:
- `src/` directory for source code
- `doc/` directory for documentation
- `test/` directory for tests

You may also want to set this python project up as a module. There are many guides online (such as [this one](https://www.digitalocean.com/community/tutorials/how-to-write-modules-in-python-3)).

This organization is needed for openjournals/joss-reviews#767
