### 2018-06-08 salabim/salabim#4 @gonsie
Details in the Readme
It would be nice for usage and/or installation instruction to be in the readme. It would also be great to have a link directly to the ReadtheDocs documentation.

These are nice-to-haves for openjournals/joss-reviews#767
