### 2018-06-08 salabim/salabim#3 @gonsie
I see that you have readthedocs documentation set up. It might be useful for that source code to exist in this repository. That will allow readthedocs to automatically generate versioned documentation.
