### 2018-06-11 salabim/salabim#5 @pspringer
Uniqueness
@salabim, the submission requirements include this:

> "Should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)"

You state how your product is similar to others, but to really motivate potential users, I recommend being very clear about what new capabilities your product offers.  You do list the capabilities, but if they are unique, say so.  Or another way of doing this might be to take a couple of popular similar products, and show areas where yours improves upon those.

This is for https://github.com/openjournals/joss-reviews/issues/767
