### 2018-06-11 salabim/salabim#6 @pspringer
Installation Notes
@salabim, Tkinter should be listed in your required packages table.

Also from what I can tell, PIL and Pillow are different things, so a sentence to clarify this difference would be helpful to a naive user.  For example, if the user only has PIL, and not Pillow, is this sufficient?

This is for https://github.com/openjournals/joss-reviews/issues/767
