### 2018-06-13 #767 @pspringer
@salabim, the documentation section on **Reading items from a file** is pretty slim.  I would recommend either adding some detail to it, or inserting a note that the Reference section has more detail.
