### 2018-06-14 salabim/salabim#7 @pspringer
Contributions
The manual section titled **Contributing to salabim** is blank.

This is for https://github.com/openjournals/joss-reviews/issues/767
