### 2018-06-14 #767 @salabim
@pspringer I have updated the section on reading files and the requirement section (re Pillow).

@pspringer I would like to say something about the unique features of salabim, but where?
