### 2018-06-15 salabim/salabim#8 @pspringer
Sample Code Discrepancy
@salabim, the explanation of simple bank example 1 in the documentation mentions a couple calls that do not show up in the sample code.  This includes self.leave(), and clerk.reactivate().  This is in regards to https://github.com/openjournals/joss-reviews/issues/767
