### 2018-06-15 #767 @pspringer
@salabim I would recommend mentioning the unique features in the README file, but let's see if @danielskatz concurs.
