### 2018-06-15 #767 @danielskatz
one or both of the following would be appropriate:
the readme, the documentation
in addition to in the paper.
