### 2018-06-15 salabim/salabim#8 @salabim
@pspringer

Thanks for pointing out the error in the documentation.
I have updated the text.

Ruud.

Ruud van der Ham
*rt.van.der.ham@gmail.com <rt.van.der.ham@gmail.com>*
