+++ README.md
@@ -6,6 +6,10 @@ Fully object oriented simulation, similar to must and Tomas.
 
 Includes process control features, resources, queues, monitors. statistical distributions.
 
-Powerful and high quality animation facilities, which can be virtually separated from the model code
+Powerful and high quality animation facilities, which can be virtually separated from the model code.
+
+SimPy3, the most used DES under Python, does not follow the same process description method and is more difficult
+to use for modellers. Also, salabim provides animation, queues, 'states', monitors for data
+collection and presentation, tracing and statistical distributions, none of which are present in SimPy3.
 
 See www.salabim.org for details. 
