### 2018-06-15 #767 @salabim
I have added the requested information to paper.md and READ.me.

Ruud van der Ham
*rt.van.der.ham@gmail.com <rt.van.der.ham@gmail.com>*
