### 2018-06-15 #767 @danielskatz
@salabim - Just looking at the paper, I think it's still too short, and also it doesn't have any references.  See http://joss.theoj.org/about#author_guidelines, and perhaps take a look at some of the accepted submissions in http://joss.theoj.org

For example, I'm not sure the paper clearly state what problems the software is designed to solve and who the target audience is?  Also, the paper could briefly explain what DES is.
