### 2018-06-15 #767 @danielskatz
@salabim - References should be specific articles or web sites.
Otherwise, I will leave it to @pspringer and @gonsie to provide more feedback on the content.
