### 2018-06-17 salabim/salabim#7 @salabim
I have deleted this section
