### 2018-06-17 salabim/salabim#5 @salabim
This is added to the paper.md file.
