### 2018-06-17 salabim/salabim#4 @salabim
I have added a link to the documentation in the README.md file.
