### 2018-06-18 29c6dbe
typo in readme link to documentation
