### 2018-06-18 salabim/salabim#2 @gonsie
Directions are here: https://help.github.com/articles/creating-releases/

This will allow users to easily download a tarball of the release. It also allows you to include release notes. Multiple commits with a release number is not really appropriate.
