### 2018-06-21 #767 @danielskatz
Also @salabim - the paper still needs to have specific references added, and cited in the body of the paper.
