### 2018-06-21 #767 @salabim
What kind of references are we talking about. I know that there some users
who are going to publish an article, but nothing concrete yet.
Is that a problem?
Otherwise, can I just cite some sample projects?

Ruud van der Ham
*rt.van.der.ham@gmail.com <rt.van.der.ham@gmail.com>*
