### 2018-06-21 #767 @danielskatz
RIght now, the references says "Several wikipedia articles on DES and simulation languages."

What are these articles?  You need to list them.
Are there academic papers about DES and the simulations languages?  If so, you should list them as well.

To pick a recent example, see https://github.com/openjournals/joss-papers/blob/joss.00773/joss.00773/10.21105.joss.00773.pdf  which has a lot of reference (likely more than you need).  The [paper](https://github.com/SMTG-UCL/galore/blob/master/paper.md) includes citations to the references, which are listed in the [bibliography file](https://github.com/SMTG-UCL/galore/blob/master/paper.bib)

This is what I'm looking for
