### 2018-06-25 #767 @pspringer
@danielskatz and @salabim I'm going through the six functional blocks I've identified to make sure all are as documented.  I should be able to finish that this week.
