### 2018-06-26 #767 @danielskatz
Add the papers to .bib file is good, but now you have to update the .md file to cite them.  Only references that are cited are built into the pdf.
