### 2018-06-27 #767 @salabim
I have updated the paper.md and paper.bib files.
Sorry for my stupid questions ...
