### 2018-06-28 #767 @danielskatz
Thanks - you can also regenerate the PDF to check it (this is not just something I can do).

Though I will do it again this time.
