### 2018-06-28 #767 @danielskatz
Those bibtex entries didn't quite work right...
