### 2018-06-28 cf75c3e
fixes to bib file
