### 2018-06-28 #767 @danielskatz
I've created 2 pull requests for the bib and the md of the paper - let's see if these fix the problems
