+++ Machine shop animated old.py
@@ -1,0 +1,270 @@
+"""
+Machine shop example
+
+Covers:
+
+- Interrupts
+- Resources: PreemptiveResource
+
+Scenario:
+  A workshop has *n* identical machines. A stream of jobs (enough to
+  keep the machines busy) arrives. Each machine breaks down
+  periodically. Repairs are carried out by one repairman. The repairman
+  has other, less important tasks to perform, too. Broken machines
+  preempt theses tasks. The repairman continues them when he is done
+  with the machine repair. The workshop works continuously.
+
+"""
+import random
+
+import salabim as sim
+
+
+RANDOM_SEED = 42
+PT_MEAN = 10.0         # Avg. processing time in minutes
+PT_SIGMA = 2.0         # Sigma of processing time
+MTTF = 300.0           # Mean time to failure in minutes
+BREAK_MEAN = 1 / MTTF  # Param. for expovariate distribution
+REPAIR_TIME = 30.0     # Time it takes to repair a machine in minutes
+JOB_DURATION = 30.0    # Duration of other jobs in minutes
+NUM_MACHINES = 10      # Number of machines in the machine shop
+WEEKS = 4              # Simulation time in weeks
+SIM_TIME = WEEKS * 7 * 24 * 60  # Simulation time in minutes
+SCALE = 10
+
+
+class MachineBarAnimate(sim.Animate):
+    def __init__(self, machine):
+        self.machine = machine
+        sim.Animate.__init__(self, rectangle0=(0, 0, 0, 0), linewidth0=0)
+
+    def rectangle(self, t):
+        if self.machine.scheduled_time() == sim.inf:
+            d = self.machine.left_to_do
+        else:
+            d = self.machine.scheduled_time() - t
+        return(
+            100, 100 + self.machine.n * 30,
+            100 + d * SCALE, 100 + self.machine.n * 30 + 20)
+
+    def fillcolor(self, t):
+        if self.machine.mode() == 'work':
+            return 'green'
+        if self.machine.mode() == 'wait':
+            return 'red'
+        if self.machine.mode() == 'repair':
+            return 'orange'
+        return ''
+
+
+class MachineTextAnimate(sim.Animate):
+    def __init__(self, machine):
+        self.machine = machine
+        sim.Animate.__init__(self,
+            x0=10, y0=100 + self.machine.n * 30, text='', anchor='sw', font='narrow', fontsize0=15)
+
+    def text(self, t):
+        return '{} {:4d}'.format(self.machine.ident, self.machine.parts_made)
+
+
+class MachineBarJobAnimate(sim.Animate):
+    def __init__(self, machine):
+        self.machine = machine
+        sim.Animate.__init__(self, rectangle0=(0, 0, 0, 0), linewidth0=0)
+
+    def rectangle(self, t):
+        d = self.machine.job_time
+        return(
+            100, 100 + self.machine.n * 30,
+            100 + d * SCALE, 100 + self.machine.n * 30 + 20)
+
+    def fillcolor(self, t):
+        if self.machine.mode() == 'work':
+            return ('green', 25)
+        if self.machine.mode() == 'wait':
+            return ('red', 25)
+        if self.machine.mode() == 'repair':
+            return ('orange', 25)
+        return ''
+
+
+class RepairBlockAnimate(sim.Animate):
+    def __init__(self, i):
+        self.i = i
+        sim.Animate.__init__(self, y0=10, rectangle0=(0, 0, 20, 20), linecolor0='white')
+
+    def x(self, t):
+        return xrepairman(self.i, t)
+
+    def rectangle(self, t):
+        if self.i == -1:
+            if repairman.claimers()[0] is None:
+                d = 0
+            else:
[…]
