### 2018-07-04 salabim/salabim#12 @pspringer
Community Guidelines
@salabim, one of the items on the reviewers' checklist is the following:

_Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_

I didn't see any statement specific to that--could you please add something that addresses this?

This is for https://github.com/openjournals/joss-reviews/issues/767
