### 2018-07-04 salabim/salabim#12 @salabim
I have added the requested information in the About section of the
documentation.

Ruud van der Ham
*rt.van.der.ham@gmail.com <rt.van.der.ham@gmail.com>*
