+++ paper.bib
@@ -1,17 +1,22 @@
 @book{Matloff:2013,
-  	Author = {{Matloff, Norm}},
-  	Booktitle = {Introduction to Discrete-Event Simulation and the SimPy Language, PDF}},
-  	Year = 2013}
+  	Author = {Matloff, Norm},
+  	Booktitle = {Introduction to Discrete-Event Simulation and the SimPy Language, PDF},
+  	Year = 2013
+  	}
 @book{Robinson:2004,
-  	Author = {{Robinson, Stewart}},
-  	Booktitle = {{Simulation – The practice of model development and use, Wiley}},
-  	Year = 2013}
+  	Author = {Robinson, Stewart},
+  	Booktitle = {Simulation – The practice of model development and use, Wiley},
+  	Year = 2013
+  	}
 @book{Page:2005,
-  	Author = {{Bernd Page}, {Wolfgang Kreutzer}},
-  	Booktitle = {{Simulating Discrete Event Systems, Shaker Verlag}},
-  	Year = 2005}
+  	Author = {Bernd Page}, {Wolfgang Kreutzer},
+  	Booktitle = {Simulating Discrete Event Systems, Shaker Verlag},
+  	Year = 2005
+  	}
 @book{Lutz:2014,
-  	Author = {{Mark Lutz}},
-  	Booktitle = {{Python Pocket Reference, O'Reilly}},
-  	Year = 2014}
+  	Author = {Mark Lutz},
+  	Booktitle = {Python Pocket Reference, O'Reilly},
+  	Year = 2014
+  	}
+
 
