### 2018-07-04 #767 @salabim
Daniel,

I have updated the paper.bib document and double checked it manually.
I still don't know how to check or generate the pdf myself. Could you be so
kind as to do that for me?

Ruud van der Ham

*rt.van.der.ham@gmail.com <rt.van.der.ham@gmail.com>*

On Thu, 28 Jun 2018 at 14:49, Daniel S. Katz <notifications@github.com>
wrote:

> I've created 2 pull requests for the bib and the md of the paper - let's
> see if these fix the problems
>
