### 2018-07-04 #767 @danielskatz
@salabim - I think you should also be able to give this command as a comment in this GitHub issue
