### 2018-07-04 #767 @danielskatz
I'm a little confused why you didn't merge my pull requests: https://github.com/salabim/salabim/pull/10 and https://github.com/salabim/salabim/pull/11
