### 2018-07-04 #767 @danielskatz
These pull requests should fix the problems.
