+++ paper.bib
@@ -9,7 +9,7 @@
   	Year = 2013
   	}
 @book{Page:2005,
-  	Author = {Bernd Page}, {Wolfgang Kreutzer},
+  	Author = {Bernd Page, Wolfgang Kreutzer},
   	Booktitle = {Simulating Discrete Event Systems, Shaker Verlag},
   	Year = 2005
   	}
