### 2018-07-04 #767 @salabim
Sorry, but I am very confused and far from an expert in GitHub.

I have now merged the pull requests (no idea what that means) and uploaded a corrected vetsion of paper.bib.

And now?

Ruud, just a simple open source software developer ...
