### 2018-07-04 #767 @danielskatz
@salabim - you might want to read through some of the [GitHub Guides](https://guides.github.com), in particular the [one on Git](https://guides.github.com/introduction/git-handbook/)
