### 2018-07-04 #767 @danielskatz
The merge of my pull request for the bib file did not work - I think you merged into the patch, rather than into the master.  Here's what I think the file should contain:

@book{Matloff:2013,
  	author = {Matloff, Norm},
  	title = {Introduction to Discrete-Event Simulation and the SimPy Language},
  	year = 2013,
  	note = {Available from \url{http://heather.cs.ucdavis.edu/~matloff/156/PLN/DESimIntro.pdf}, Accessed 2018-07-04}
  	}
@book{Robinson:2004,
  	author = {Robinson, Stewart},
  	title = {Simulation – The practice of model development and use},
 	publisher={Wiley},
  	year = 2013
  	}
@book{Page:2005,
  	author = {Bernd Page and Wolfgang Kreutzer},
  	title = {Simulating Discrete Event Systems},
 	publisher={Shaker Verlag},
  	year = 2005
  	}
@book{Lutz:2014,
  	author = {Mark Lutz},
  	title = {Python Pocket Reference},
 	publisher = {O'Reilly},
  	year = 2014
  	}
