+++ paper.bib
@@ -1,20 +1,24 @@
 @book{Matloff:2013,
-  	Author = {Matloff, Norm},
-  	Booktitle = {Introduction to Discrete-Event Simulation and the SimPy Language, PDF},
-  	Year = 2013
-  	}
+author = {Matloff, Norm},
+title = {Introduction to Discrete-Event Simulation and the SimPy Language},
+year = 2013,
+note = {Available from \url{http://heather.cs.ucdavis.edu/~matloff/156/PLN/DESimIntro.pdf}, Accessed 2018-07-04}
+}
 @book{Robinson:2004,
-  	Author = {Robinson, Stewart},
-  	Booktitle = {Simulation – The practice of model development and use, Wiley},
-  	Year = 2013
-  	}
+author = {Robinson, Stewart},
+title = {Simulation – The practice of model development and use},
+publisher={Wiley},
+year = 2013
+}
 @book{Page:2005,
-  	Author = {Bernd Page, Wolfgang Kreutzer},
-  	Booktitle = {Simulating Discrete Event Systems, Shaker Verlag},
-  	Year = 2005
-  	}
+author = {Bernd Page and Wolfgang Kreutzer},
+title = {Simulating Discrete Event Systems},
+publisher={Shaker Verlag},
+year = 2005
+}
 @book{Lutz:2014,
-  	Author = {Mark Lutz},
-  	Booktitle = {Python Pocket Reference, O'Reilly},
-  	Year = 2014
-  	}
+author = {Mark Lutz},
+title = {Python Pocket Reference},
+publisher = {O'Reilly},
+year = 2014
+}
