### 2018-07-04 #767 @salabim
I have now put a new paper.bib, with your contents on the master branch.

Does it work now?
