### 2018-07-04 #767 @danielskatz
close

let's try changing the first entry to

```
@book{Matloff:2013,
  	author = {Matloff, Norm},
  	title = {Introduction to Discrete-Event Simulation and the Sim{P}y Language},
  	year = 2013,
  	url = {http://heather.cs.ucdavis.edu/~matloff/156/PLN/DESimIntro.pdf}
  	}
```

Then you should be able to go to the GitHub Issue where the review is taking place (this one, https://github.com/openjournals/joss-reviews/issues/767), and type "@whedon generate pdf" a new comment to generate a new PDF.
