+++ paper.bib
@@ -1,9 +1,9 @@
 @book{Matloff:2013,
-author = {Matloff, Norm},
-title = {Introduction to Discrete-Event Simulation and the SimPy Language},
-year = 2013,
-note = {Available from \url{http://heather.cs.ucdavis.edu/~matloff/156/PLN/DESimIntro.pdf}, Accessed 2018-07-04}
-}
+  	author = {Matloff, Norm},
+  	title = {Introduction to Discrete-Event Simulation and the Sim{P}y Language},
+  	year = 2013,
+  	url = {http://heather.cs.ucdavis.edu/~matloff/156/PLN/DESimIntro.pdf}
+  	}
 @book{Robinson:2004,
 author = {Robinson, Stewart},
 title = {Simulation – The practice of model development and use},
