### 2018-07-06 #767 @pspringer
@salabim, @danielskatz, the References section in paper.md is blank.  Also, where is the link to the pdf version of the paper, other than whedon's "Check article proof"?
