### 2018-07-06 #767 @danielskatz
@pspringer - don't worry about the lack of references in the .md file, they are built automatically in the PDF from the .bib file.  And the latest "Check article proof" is the most recent version of the pdf version.
