### 2018-07-07 #767 @pspringer
@danielskatz, @salabim, I've checked everything off on the list, it looks good.
