### 2018-07-07 #767 @gonsie
@salabim do any of the paper references have DOIs?
