+++ paper.bib
@@ -1,9 +1,9 @@
 @book{Matloff:2013,
-  	author = {Matloff, Norm},
-  	title = {Introduction to Discrete-Event Simulation and the Sim{P}y Language},
-  	year = 2013,
-  	url = {http://heather.cs.ucdavis.edu/~matloff/156/PLN/DESimIntro.pdf}
-  	}
+author = {Matloff, Norm},
+title = {Introduction to Discrete-Event Simulation and the Sim{P}y Language},
+year = 2013,
+url = {http://heather.cs.ucdavis.edu/~matloff/156/PLN/DESimIntro.pdf}
+}
 @book{Robinson:2004,
 author = {Robinson, Stewart},
 title = {Simulation – The practice of model development and use},
@@ -15,6 +15,7 @@ author = {Bernd Page and Wolfgang Kreutzer},
 title = {Simulating Discrete Event Systems},
 publisher={Shaker Verlag},
 year = 2005
+doi = {10.1065/espr2006.09.348}
 }
 @book{Lutz:2014,
 author = {Mark Lutz},
