+++ paper.bib
@@ -14,7 +14,7 @@ year = 2013
 author = {Bernd Page and Wolfgang Kreutzer},
 title = {Simulating Discrete Event Systems},
 publisher={Shaker Verlag},
-year = 2005
+year = 2005,
 doi = {10.1065/espr2006.09.348}
 }
 @book{Lutz:2014,
