### 2018-07-07 #767 @salabim
@gonsie
I have found one DOI and added it to the references (paper.bib).
