### 2018-07-07 #767 @danielskatz
@gonsie - Can you go ahead and check off the last box?  The references are books which generally don't have DOIs.
