### 2018-07-07 #767 @gonsie
@danielskatz done.
