### 2018-07-09 #767 @salabim
@arfon
The DOI of the figshare repository is
10.6084/m9.figshare.6791129
